+++ packages/serverless-workflow-standalone-editor/resources/swf/index.html
