+++ packages/serverless-workflow-standalone-editor/resources/swf/swfCombinedEditorEnvelopeIndex.template
