+++ packages/serverless-workflow-standalone-editor/resources/swf/swfDiagramEditorEnvelopeIndex.template
