+++ packages/serverless-workflow-standalone-editor/resources/swf/swfMermaidViewerEnvelopeIndex.template
