+++ packages/serverless-workflow-standalone-editor/resources/swf/swfTextEditorEnvelopeIndex.template
