+++ packages/serverless-workflow-standalone-editor/src/common/EditorResources.ts
