+++ packages/serverless-workflow-standalone-editor/src/envelope/SwfCombinedEditorEnvelopeApp.ts
