+++ packages/serverless-workflow-standalone-editor/src/envelope/SwfDiagramEditorEnvelopeApp.ts
