+++ packages/serverless-workflow-standalone-editor/src/envelope/SwfMermaidViewerEnvelopeApp.ts
