+++ packages/serverless-workflow-standalone-editor/src/envelope/SwfTextEditorEnvelopeApp.ts
