+++ packages/serverless-workflow-standalone-editor/src/global.d.ts
