+++ packages/serverless-workflow-standalone-editor/src/swf/resources/SwfCombinedEditorResources.ts
