+++ packages/serverless-workflow-standalone-editor/src/swf/resources/SwfDiagramEditorResources.ts
