+++ packages/serverless-workflow-standalone-editor/src/swf/resources/SwfMermaidViewerResources.ts
