+++ packages/serverless-workflow-standalone-editor/src/swf/resources/SwfTextEditorResources.ts
