+++ packages/serverless-workflow-standalone-editor/webpack.build-resources.config.js
