+++ packages/serverless-workflow-standalone-editor/webpack.editor-resources.config.js
