+++ packages/serverless-workflow-standalone-editor/webpack.package-resources.config.js
